mvp 1차 작업진행
