+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".view.MainActivity">
+        <activity android:name="com.example.kakaosearcher.kakaoaddress.view.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
