+++ app/src/main/java/com/example/kakaosearcher/kakaoaddress/model/reqmodel/ReqAddressModel.kt
@@ -1,0 +1,1 @@
+package com.example.kakaosearcher.kakaoaddress.model.reqmodel
