+++ app/src/main/java/com/example/kakaosearcher/kakaoaddress/presenter/AddressContract.kt
@@ -1,0 +1,15 @@
+package com.example.kakaosearcher.kakaoaddress.presenter
+
+import com.example.kakaosearcher.kakaoaddress.model.resmodel.AddressModel
+
+interface AddressContract {
+
+    interface View {
+        fun updateAddressList(addressList: List<AddressModel>)
+        fun showErrorMesage(msg: String)
+    }
+
+    interface Presenter {
+        fun searchAddress(query: String)
+    }
+}
