+++ app/src/main/java/com/example/kakaosearcher/kakaoaddress/repository/AddressRepository.kt
@@ -1,13 +1,13 @@
-package com.example.kakaosearcher.controller
+package com.example.kakaosearcher.kakaoaddress.repository
 
-import com.example.kakaosearcher.network.model.resmodel.dto.AddressDto
+import com.example.kakaosearcher.kakaoaddress.model.resmodel.dto.AddressDto
 import com.example.kakaosearcher.network.retrofit.RetrofitCallBack
 import com.example.kakaosearcher.network.retrofit.RetrofitListener
 import com.example.kakaosearcher.network.service.AddressService
 
-class AddressController {
+class AddressRepository {
 
-    fun requestAddress(
+    fun getAddress(
         query: String?,
         listener: RetrofitListener<AddressDto>
     ) {
