--- app/src/main/java/com/example/kakaosearcher/network/model/reqmodel/ReqAddressModel.kt
@@ -1,1 +1,0 @@
-package com.example.kakaosearcher.network.model.reqmodel
