+++ app/src/main/res/layout/fragment_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.MainActivity">
+    tools:context=".kakaoaddress.view.MainActivity">
 
     <EditText
         android:id="@+id/et_fragment_main_input_text"
